+++ src/datagrid/column-service/column-types/multiselect-renderers.jsx
@@ -1,6 +1,10 @@
 import React from 'react';
+import { fromJS, List } from 'immutable';
+import { OverlayTrigger, Tooltip } from 'react-bootstrap';
 import { FloatingSelectPortal } from '@opuscapita/react-floating-select';
 import Checkbox from '@opuscapita/react-checkbox';
+
+// TODO: cellCreate, cellEdit renderers -> multiselect dropdown as cellFilter has
 
 const MultiValueContainer = (label, optionsCount) => properties => (properties.selectProps.value
   && properties.selectProps.value.findIndex(
@@ -10,6 +14,66 @@ const MultiValueContainer = (label, optionsCount) => properties => (properties.s
   ) : null);
 
 export default {
+  valRender(
+    col,
+    column,
+    rowIndex,
+    gridId,
+    selectOptions = [],
+    valueRender,
+  ) {
+    return valueRender(rowIndex, (values) => {
+      // normalize values as Immutable List
+      let vals = values;
+      if (Array.isArray(values)) {
+        vals = fromJS(values);
+      }
+      if (!List.isList(vals)) {
+        vals = fromJS([values]);
+      }
+      // No items return empty string
+      if (!vals.size) return '';
+      // If only one is selected render value normally
+      if (vals.size === 1) {
+        const option = selectOptions.find(o => o.value === vals.get(0));
+        return option ? option.label : null;
+      }
+      // Multiple values selected, render '3 selected'
+      const selectedText = (col.translations && col.translations.itemsSelected) || 'selected';
+      const valueText = (
+        <span>
+          {`${vals.size} ${selectedText}`}
+        </span>
+      );
+      // without tooltip
+      if (col.isMultiselectTooltipDisabled) {
+        return valueText;
+      }
+      // with tooltip
+      return (
+        <OverlayTrigger
+          placement="top"
+          overlay={(
+            <Tooltip id={`ocDatagridMultiselectTooltip-${gridId}-${column.columnKey}-${rowIndex}`}>
+              <div style={{ textAlign: 'left' }}>
+                {selectOptions.filter(o => vals.includes(o.value)).map(v => (
+                  <React.Fragment key={v.value}>
+                    <span>
+                      {v.label}
+                    </span>
+                    <br />
+                  </React.Fragment>
+                ))}
+              </div>
+            </Tooltip>
+          )}
+          delay={500}
+        >
+          {valueText}
+        </OverlayTrigger>
+      );
+    });
+  },
   cellFilter(
     col,
     column,